:white_check_mark: added grpc getallocation test
